requirements.txt
torch
torchvision
Pillow
scikit-learn
numpy
pandas
matplotlib
PyYAML
timm
pytorch-lightning
tqdm
opencv-python
mlflow

